# apps/server/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ashyq Bilim API",
    "description": "University LMS / MOOC platform API (v2, Rust rewrite).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/course-updates/{id}": {
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete an announcement (course write access).",
        "operationId": "delete_course_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course update id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseUpdateId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Edit an announcement (course write access).",
        "operationId": "edit_course_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course update id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseUpdateId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCourseUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseUpdate"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Newest-first course listing: public courses plus the caller's own\n(readers with `course:read:all` see everything).",
        "operationId": "list_courses",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "next_cursor from the previous page",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100 (default 20)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of courses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursePage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create a course (requires `course:create:platform`).",
        "operationId": "create_course",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/courses/{id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "One course (404 for private courses the caller cannot see).",
        "operationId": "get_course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete a course and everything under it (cascades).",
        "operationId": "delete_course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No delete access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Partial update (creator with `course:update:own` or platform updaters).",
        "operationId": "update_course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/courses/{id}/lifecycle": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Publish/unpublish (legacy lifecycle semantics).",
        "operationId": "course_lifecycle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseLifecycleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Visibility changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/courses/{id}/updates": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Course announcements, newest first (read follows course visibility).",
        "operationId": "list_course_updates",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Announcements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseUpdate"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post an announcement (course write access).",
        "operationId": "create_course_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseUpdate"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Course": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "about",
          "tags",
          "public",
          "open_to_contributors",
          "created_at_unix",
          "updated_at_unix"
        ],
        "properties": {
          "about": {
            "type": "string"
          },
          "created_at_unix": {
            "type": "integer",
            "format": "int64"
          },
          "creator_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "description": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/CourseId"
          },
          "name": {
            "type": "string"
          },
          "open_to_contributors": {
            "type": "boolean"
          },
          "public": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updated_at_unix": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CourseId": {
        "type": "string",
        "format": "uuid"
      },
      "CourseLifecycleRequest": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "`publish` or `unpublish`."
          }
        },
        "additionalProperties": false
      },
      "CoursePage": {
        "type": "object",
        "description": "Keyset page (ARCHITECTURE §6): pass `next_cursor` back as `cursor`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "next_cursor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CourseId"
              }
            ]
          }
        }
      },
      "CourseUpdate": {
        "type": "object",
        "description": "One announcement in the course changelog feed.",
        "required": [
          "id",
          "course_id",
          "title",
          "content",
          "created_at_unix",
          "updated_at_unix"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "course_id": {
            "$ref": "#/components/schemas/CourseId"
          },
          "created_at_unix": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "$ref": "#/components/schemas/CourseUpdateId"
          },
          "title": {
            "type": "string"
          },
          "updated_at_unix": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CourseUpdateId": {
        "type": "string",
        "format": "uuid"
      },
...
      "CreateCourseRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "about": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "CreateCourseUpdateRequest": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "EditCourseUpdateRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "ErrorCode": {
        "type": "string",
        "description": "Stable, closed set of machine-readable error codes.",
        "enum": [
          "internal",
          "not-found",
          "forbidden",
          "unauthenticated",
          "conflict",
          "validation-failed",
          "rate-limited",
          "payload-too-large",
          "service-unavailable",
          "invalid-credentials",
          "mfa-required",
          "session-expired",
          "csrf-rejected",
          "account-disabled",
          "google-oauth-expired",
          "invalid-totp-code"
        ]
      },
      "FieldError": {
        "type": "object",
        "description": "A single field-level validation failure. `code` is a stable machine key the\nfrontend translates (e.g. `required`, `too-long`); `message` is English.",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "The wire shape of every error response.",
        "required": [
          "type",
          "status",
          "code",
          "title"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine code — the frontend's i18n key."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "field_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id; also present as the `x-request-id` response header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Stable documentation URI for this error class."
          }
        }
      },
...
      "UpdateCourseRequest": {
        "type": "object",
        "properties": {
          "about": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "open_to_contributors": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "UserId": {
        "type": "string",
        "format": "uuid"
      },
...
    }
  },
...
}